# Minor code quality/consistency improvements