--- a/src/Esprima/EsprimaExceptionHelper.cs
+++ b/src/Esprima/EsprimaExceptionHelper.cs
@@ -7,17 +7,17 @@ namespace Esprima;
 /// </remarks>
 internal static class EsprimaExceptionHelper
 {
-    [DoesNotReturn]
-    public static void ThrowIndexOutOfRangeException()
-    {
-        throw new IndexOutOfRangeException();
-    }
+    //[DoesNotReturn]
+    //public static void ThrowIndexOutOfRangeException()
+    //{
+    //    throw new IndexOutOfRangeException();
+    //}
 
-    [DoesNotReturn]
-    public static T ThrowArgumentOutOfRangeException<T>(string paramName, object actualValue, string? message = null)
-    {
-        throw new ArgumentOutOfRangeException(paramName, actualValue, message);
-    }
+    //[DoesNotReturn]
+    //public static T ThrowArgumentOutOfRangeException<T>(string paramName, object actualValue, string? message = null)
+    //{
+    //    throw new ArgumentOutOfRangeException(paramName, actualValue, message);
+    //}
 
     [DoesNotReturn]
     public static void ThrowArgumentOutOfRangeException(string paramName, object actualValue, string? message = null)
@@ -31,21 +31,27 @@ internal static class EsprimaExceptionHelper
         throw new InvalidOperationException(message);
     }
 
-    [DoesNotReturn]
-    public static void ThrowInvalidOperationException(string? message = null)
-    {
-        throw new InvalidOperationException(message);
-    }
+    //[DoesNotReturn]
+    //public static void ThrowInvalidOperationException(string? message = null)
+    //{
+    //    throw new InvalidOperationException(message);
+    //}
 
-    [DoesNotReturn]
-    public static void ThrowArgumentNullException(string message)
-    {
-        throw new ArgumentNullException(message);
-    }
+    //[DoesNotReturn]
+    //public static void ThrowArgumentNullException(string message)
+    //{
+    //    throw new ArgumentNullException(message);
+    //}
 
     [DoesNotReturn]
     public static T ThrowArgumentNullException<T>(string message)
     {
         throw new ArgumentNullException(message);
+    }
+
+    [DoesNotReturn]
+    public static T ThrowFormatException<T>(string message)
+    {
+        throw new FormatException(message);
     }
 }