… src/Esprima/Exception.cs
… src/Esprima/Loc.cs
… src/Esprima/ParseError.cs
… src/Esprima/Position.cs
… src/Esprima/Range.cs
… src/Esprima/Scanner.cs
… src/Esprima/StringPool.cs
… src/Esprima/Utils/JsonTextWriter.cs
… test/Esprima.Tests/AstToJavascriptTests.cs