Some additional cleanup (but no public APIs were harmed in the process 😄)

Also did a few micro-optimizations along the way. I'm not sure about [this one](https://github.com/sebastienros/esprima-dotnet/commit/b0fb4229fedf18edf5766df11b07313e2087ec44). It works fine, I just don't know if you're ok with such techniques or consider it too hacky?