@adams85 2022-08-24
> Can you add before/after diffs with BenchmarkDotNet.ResultDiff kinda way for FileParsing so we know nothing went south, otherwise look like nice improvements!

Sure:
## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|18.524 ms|0.2234 ms|687.5000|343.7500|-|4,314 KB|
| **New** |	|	| **18.329 ms (-1%)** | **0.0681 ms** | **718.7500 (+5%)** | **343.7500 (0%)** | **-** | **4,314 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|2.473 ms|0.0219 ms|140.6250|27.3438|-|676 KB|
| **New** |	|	| **2.509 ms (+1%)** | **0.0280 ms** | **140.6250 (0%)** | **27.3438 (0%)** | **-** | **676 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|14.417 ms|0.0676 ms|625.0000|312.5000|-|3,765 KB|
| **New** |	|	| **14.596 ms (+1%)** | **0.1079 ms** | **640.6250 (+3%)** | **312.5000 (0%)** | **-** | **3,765 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|23.022 ms|0.3184 ms|1000.0000|468.7500|31.2500|5,800 KB|
| **New** |	|	| **23.129 ms (0%)** | **0.0653 ms** | **1000.0000 (0%)** | **468.7500 (0%)** | **31.2500 (0%)** | **5,800 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|11.746 ms|0.0836 ms|500.0000|250.0000|-|3,090 KB|
| **New** |	|	| **11.812 ms (+1%)** | **0.0537 ms** | **500.0000 (0%)** | **250.0000 (0%)** | **-** | **3,090 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|2.101 ms|0.0262 ms|117.1875|42.9688|-|571 KB|
| **New** |	|	| **2.098 ms (0%)** | **0.0225 ms** | **117.1875 (0%)** | **42.9688 (0%)** | **-** | **571 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|10.815 ms|0.0654 ms|468.7500|234.3750|-|2,857 KB|
| **New** |	|	| **10.810 ms (0%)** | **0.0654 ms** | **468.7500 (0%)** | **234.3750 (0%)** | **-** | **2,857 KB (0%)** |

As expected, no significant changes.